requirements.txt
fastapi
uvicorn
jinja2
python-multipart
transformers
torch
sentence-transformers
faiss-cpu
redis
slowapi
detoxify
loguru
pydantic